requirements.txt
# List required packages in this file, one per line.
airtable-python-wrapper
ruamel.yaml
nameparser
fire
